Update user input (only support legacy C inputs)

+++ jupyter_cpp_kernel/kernel.py
@@ -9,7 +9,8 @@ import os
 import os.path as path
 
 class RealTimeSubprocess(subprocess.Popen):
-    def __init__(self, cmd, write_to_stdout, write_to_stderr):
+    inputRequest = "<inputRequest>"
+    def __init__(self, cmd, write_to_stdout, write_to_stderr, read_from_stdin):
         """
         cmd: the command to execute
         write_to_stdout: a callable that will be called with chunks of data from stdout
@@ -17,8 +18,9 @@ class RealTimeSubprocess(subprocess.Popen):
         """
         self._write_to_stdout = write_to_stdout
         self._write_to_stderr = write_to_stderr
+        self._read_from_stdin = read_from_stdin
 
-        super().__init__(cmd, stdin=subprocess.PIPE, stdout=subprocess.PIPE, stderr=subprocess.PIPE, bufsize=0)
+        super().__init__(cmd, stdout=subprocess.PIPE, stderr=subprocess.PIPE, stdin=subprocess.PIPE, bufsize=0)
 
         self._stdout_queue = Queue()
         self._stdout_thread = Thread(target=RealTimeSubprocess._enqueue_output, args=(self.stdout, self._stdout_queue))
@@ -32,13 +34,19 @@ class RealTimeSubprocess(subprocess.Popen):
 
     @staticmethod
     def _enqueue_output(stream, queue):
-        # Add chunks of data from a stream to a queue until the stream is empty.
+        """
+        Add chunks of data from a stream to a queue until the stream is empty.
+        """
         for line in iter(lambda: stream.read(4096), b''):
             queue.put(line)
         stream.close()
 
     def write_contents(self):
-        # Write the available content from stdin and stderr where specified when the instance was created
+        """
+        Write the available content from stdin and stderr where specified when the instance was created
+        :return:
+        """
+
         def read_all_from_queue(queue):
             res = b''
             size = queue.qsize()
@@ -47,12 +55,25 @@ class RealTimeSubprocess(subprocess.Popen):
                 size -= 1
             return res
 
-        stdout_contents = read_all_from_queue(self._stdout_queue)
-        if stdout_contents:
-            self._write_to_stdout(stdout_contents)
         stderr_contents = read_all_from_queue(self._stderr_queue)
         if stderr_contents:
-            self._write_to_stderr(stderr_contents)
+            self._write_to_stderr(stderr_contents.decode())
+
+        stdout_contents = read_all_from_queue(self._stdout_queue)
+        if stdout_contents:
+            contents = stdout_contents.decode()
+            start = contents.find(self.__class__.inputRequest)
+            if(start >= 0):
+                contents = contents.replace(self.__class__.inputRequest, '')
+                if(len(contents) > 0):
+                    self._write_to_stdout(contents)
+                readLine = ""
+                while(len(readLine) == 0):
+                    readLine = self._read_from_stdin()
+                readLine += "\n"
+                self.stdin.write(readLine.encode())
+            else:
+                self._write_to_stdout(contents)
 
 
 class CPPKernel(Kernel):
@@ -77,26 +98,40 @@ class CPPKernel(Kernel):
     
     banner = "C++ 14 kernel for Jupyter\n\nGCC version:\n" + gcc_version + "\n\nG++ version:\n" + gpp_version + "\n\n"
 
+    main_head = "#include <iostream>\n" \
+            "int main(){\n"
+
+    main_foot = "\nreturn 0;\n}"
+
     def __init__(self, *args, **kwargs):
         super(CPPKernel, self).__init__(*args, **kwargs)
+        self._allow_stdin = True
+        self.readOnlyFileSystem = False
+        self.bufferedOutput = True
+        self.linkMaths = True # always link math library
+        self.wAll = True # show all warnings by default
+        self.wError = False # but keep comipiling for warnings
+        self.standard = "c++14" # default standard if none is specified
         self.files = []
         mastertemp = tempfile.mkstemp(suffix='.out')
         os.close(mastertemp[0])
         self.master_path = mastertemp[1]
-        filepath = path.join(path.dirname(path.realpath(__file__)), 'resources', 'master.cpp')
+        self.resDir = path.join(path.dirname(path.realpath(__file__)), 'resources')
+        filepath = path.join(self.resDir, 'master.cpp')
         subprocess.call(['g++', filepath, '-std=c++14', '-rdynamic', '-ldl', '-o', self.master_path])
 
     def cleanup_files(self):
-        # Remove all the temporary files created by the kernel
+        """Remove all the temporary files created by the kernel"""
+        # keep the list of files create in case there is an exception
+        # before they can be deleted as usual
         for file in self.files:
-            os.remove(file)
+            if(os.path.exists(file)):
+                os.remove(file)
         os.remove(self.master_path)
 
     def new_temp_file(self, **kwargs):
-        """
-        Create a new temp file to be deleted when the kernel shuts down
-        We don't want the file to be deleted when closed, but only when the kernel stops
-        """
+        """Create a new temp file to be deleted when the kernel shuts down"""
+        # We don't want the file to be deleted when closed, but only when the kernel stops
         kwargs['delete'] = False
         kwargs['mode'] = 'w'
         file = tempfile.NamedTemporaryFile(**kwargs)
@@ -109,75 +144,95 @@ class CPPKernel(Kernel):
     def _write_to_stderr(self, contents):
         self.send_response(self.iopub_socket, 'stream', {'name': 'stderr', 'text': contents})
 
-    def create_jupyter_subprocess(self, cmd, input_text=""):
+    def _read_from_stdin(self):
+        return self.raw_input()
+
+    def create_jupyter_subprocess(self, cmd):
         return RealTimeSubprocess(cmd,
-                                  lambda contents: self._write_to_stdout(contents.decode()),
-                                  lambda contents: self._write_to_stderr(contents.decode()))
+                                  self._write_to_stdout,
+                                  self._write_to_stderr,
+                                  self._read_from_stdin)
 
     def compile_with_gpp(self, source_filename, binary_filename, cflags=None, ldflags=None):
+        cflags = ['-pedantic', '-fPIC', '-shared', '-rdynamic'] + cflags
+        if self.linkMaths:
+            cflags = cflags + ['-lm']
+        if self.wError:
+            cflags = cflags + ['-Werror']
+        if self.wAll:
+            cflags = cflags + ['-Wall']
+        if self.readOnlyFileSystem:
+            cflags = ['-DREAD_ONLY_FILE_SYSTEM'] + cflags
+        if self.bufferedOutput:
+            cflags = ['-DBUFFERED_OUTPUT'] + cflags
         args = ['g++', source_filename] + cflags + ['-o', binary_filename] + ldflags
-        return self.create_jupyter_subprocess(cmd=args)
+        return self.create_jupyter_subprocess(args)
 
     def _filter_magics(self, code):
 
         magics = {'cflags': [],
                   'ldflags': [],
-                  'stdin': "",
-                  'stdout': "",
-                  'memtotalnoterm': "",
-                  'memtotal': "",
-                  'memaux': "",
-                  'memexpect': "",
-                  'test_script': [],
                   'args': []}
+
+        actualCode = ''
 
         for line in code.splitlines():
             if line.startswith('//%'):
-                try:
-                    key, value = line[3:].split(":", 2)
-                    key = key.strip().lower()
-                except ValueError:
-                    # Don't want kernel to fail because an error was made in source code comment.
+                magicSplit = line[3:].split(":", 2)
+                if(len(magicSplit) < 2):
+                    self._write_to_stderr("[C++ kernel] Magic line starting with '//%' is missing a semicolon, ignoring.")
                     continue
+
+                key, value = magicSplit
+                key = key.strip().lower()
 
                 if key in ['ldflags', 'cflags']:
                     for flag in value.split():
                         magics[key] += [flag]
-                elif key == "stdin":
-                    # Could probably be a match instead,
-                    # but this is easier for now.
-                    # Note: this is very basic, string escapes and the like
-                    #   aren't included.
-                    for stringContents in re.findall(r'\s*"([^"]*)"', value):
-                        magics['stdin'] += stringContents + "\n"
-                elif key == "stdout":
-                    # Could probably be a match instead,
-                    # but this is easier for now.
-                    # Note: this is very basic, string escapes and the like
-                    #   aren't included.
-                    for stringContents in re.findall(r'\s*"([^"]*)"', value):
-                        magics['stdout'] += stringContents + "\n"
-                elif key == "test_script":
-                    magics['test_script'].append(value)
-                elif key == "memtotalnoterm":
-                    magics['memtotalnoterm'] += "If you only allocated " + value + " bytes, you have likely left off space for string termination.\n"
-                elif key == "memtotal":
-                    magics['memtotal'] += "If you only allocated " + value + " bytes, you may not have allocated space for the array, which may or may not be what you wanted to do.\n"
-                elif key == "memaux":
-                    magics['memaux'] += "If you only allocated " + value + " bytes, you may not have allocated space for each string in the array, only the pointers to each string.\n"
-                elif key == "memexpect":
-                    magics['memexpect'] += "You should expect to use about " + value + " bytes, you can use valgrind and gdb to check that you've got this right in next week's workshop.\n"
                 elif key == "args":
                     # Split arguments respecting quotes
                     for argument in re.findall(r'(?:[^\s,"]|"(?:\\.|[^"])*")+', value):
                         magics['args'] += [argument.strip('"')]
 
-        return magics
+                # always add empty line, so line numbers don't change
+                actualCode += '\n'
+
+            # keep lines which did not contain magics
+            else:
+                actualCode += line + '\n'
+
+        # add default standard if cflags does not contain one
+        if not any(item.startswith('-std=') for item in magics["cflags"]):
+            magics["cflags"] += ["-std=" + self.standard]
+
+        return magics, actualCode
+
+    # check whether int main() is specified, if not add it around the code
+    # also add common magics like -lm
+    def _add_main(self, magics, code):
+        # remove comments
+        tmpCode = re.sub(r"//.*", "", code)
+        tmpCode = re.sub(r"/\*.*?\*/", "", tmpCode, flags=re.M|re.S)
+
+        x = re.search(r"int\s+main\s*\(", tmpCode)
+
+        if not x:
+            code = self.main_head + code + self.main_foot
+            magics['cflags'] += ['-lm']
+
+        return magics, code
 
     def do_execute(self, code, silent, store_history=True,
-                   user_expressions=None, allow_stdin=False):
+                   user_expressions=None, allow_stdin=True):
 
-        magics = self._filter_magics(code)
+        magics, code = self._filter_magics(code)
+
+        magics, code = self._add_main(magics, code)
+
+        # replace stdio with wrapped version
+        headerDir = "\"" + self.resDir + "/stdio_wrap.h" + "\""
+        code = code.replace("<stdio.h>", headerDir)
+        code = code.replace("\"stdio.h\"", headerDir)
 
         with self.new_temp_file(suffix='.cpp') as source_file:
             source_file.write(code)
@@ -188,49 +243,35 @@ class CPPKernel(Kernel):
                     p.write_contents()
                 p.write_contents()
                 if p.returncode != 0:  # Compilation failed
-                    self._write_to_stderr("Kernel location: " + os.path.dirname(os.path.realpath(__file__)) + "\n")
                     self._write_to_stderr(
-                            "[C++ 14 kernel] G++ exited with code {}, the code will not be executed".format(
+                            "[C++ 14 kernel] G++ exited with code {}, the executable will not be executed".format(
                                     p.returncode))
+
+                    # delete source files before exit
+                    os.remove(source_file.name)
+                    os.remove(binary_file.name)
+
                     return {'status': 'ok', 'execution_count': self.execution_count, 'payload': [],
                             'user_expressions': {}}
 
-        p = self.create_jupyter_subprocess(cmd=([binary_file.name] + magics['args']))
-        if magics['stdin'] != "":
-            self._write_to_stderr(("input: " + magics['stdin']))
-        if magics['stdout'] != "":
-            self._write_to_stderr(("expected output: " + magics['stdout']))
-        if magics['memtotalnoterm'] != "" or magics['memtotal'] != "" or magics['memaux'] != "" or magics['memexpect'] != "":
-            self._write_to_stderr("Some memory hints which you might like to verify for this question:\n")
-        if magics['memtotalnoterm'] != "":
-            self._write_to_stderr(magics['memtotalnoterm'])
-        if magics['memtotal'] != "":
-            self._write_to_stderr(magics['memtotal'])
-        if magics['memaux'] != "":
-            self._write_to_stderr(magics['memaux'])
-        if magics['memexpect'] != "":
-            self._write_to_stderr(magics['memexpect'])
-        p.stdin.write(magics['stdin'].encode(encoding="utf-8", errors="strict"))
-        p.stdin.close()
+        p = self.create_jupyter_subprocess([self.master_path, binary_file.name] + magics['args'])
         while p.poll() is None:
             p.write_contents()
+
+        # wait for threads to finish, so output is always shown
+        p._stdout_thread.join()
+        p._stderr_thread.join()
+
         p.write_contents()
 
+        # now remove the files we have just created
+        os.remove(source_file.name)
+        os.remove(binary_file.name)
+
         if p.returncode != 0:
-            self._write_to_stderr("[C++ kernel] Executable exited with code {}".format(p.returncode))
-
-        if len(magics['test_script']) > 0:
-            for item in magics['test_script']:
-                self._write_to_stderr("Testing with script: " + item + ".\n")
-                subproc_commands = [['chmod'] + ['777'] + [item]] + [['/bin/bash'] + [item] + [binary_file.name]]
-                for subproc_command in subproc_commands:
-                    self._write_to_stderr("Command: " + str(subproc_command) + '\n')
-                    test_process = self.create_jupyter_subprocess(cmd=subproc_command)
-                    while test_process.poll() is None:
-                        test_process.write_contents()
-                    test_process.write_contents()
-
+            self._write_to_stderr("[C++ 14 kernel] Executable exited with code {}".format(p.returncode))
         return {'status': 'ok', 'execution_count': self.execution_count, 'payload': [], 'user_expressions': {}}
 
     def do_shutdown(self, restart):
+        """Cleanup the created source code files and executables when shutting down the kernel"""
         self.cleanup_files()
